test: run the E2E suite against the Docker image
`make test-e2e` exercises `next build && next start`, not the standalone
build we publish, so a fault living only in the packaged image — a file the
tracer didn't copy, a path that resolves differently under /data, the
container's UTC clock — could reach a release with every tier green.

Adds `make test-e2e-docker`, which builds the image, runs the suite against a
container, and slots into the release checklist between creating the tag and
pushing it, where a failure still costs nothing. Seeding stays on the host and
reaches the container through the /data bind mount, so no seeding code has to
ship in the image.
